[!!!][TASK] Migrate to doctrine and drop redirects
This patch migrates the final parts of the extension
(the administration module) to doctrine based
querries.

Also the redirect functionality of realurl is dropped
entirely in favor of the redirect functionality of the
TYPO3 core. However, the database table containing the
redirects is still in place to enable a migration.